refactored config access module
