+++ airflow_logic/dags/osm_dags.py
@@ -5,7 +5,7 @@ from airflow_logic.dags.dags_utils.dag_configuration import (
     get_default_arguments,
     get_schedule_interval,
 )
-from dags_utils.utils.config_parser import config
+from config_access.config_parser import config
 from dags_utils.utils.osm_dags_filler import fill_osm_dag
 
 # Defaults which can be overridden if needed
