+++ auxiliary_modules/config_access/config_parser.py
@@ -2,9 +2,8 @@ import os
 from pathlib import Path
 from typing import List
 
+from config_access.fallback_dict import FallbackDict
 import yaml
-
-from utils.fallback_dict import FallbackDict
 
 
 class Config:
